requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
mysqlclient==2.2.0

# CORS Headers
django-cors-headers==4.3.0

# Environment variables
python-decouple==3.8

# API Documentation
drf-yasg==1.21.7

# Filtering and Pagination
django-filter==23.3

# Timezone support
pytz==2023.3

# Production server
gunicorn==21.2.0

# Redis for caching (optional)
redis==5.0.1
django-redis==5.4.0
